refactor(test): split session-phase-machine into 3 files for shard balance (#233)
The `test/session-phase-machine.test.ts` monolith (25 flat `it` blocks under
one `describe`, 88.7 s on CI) both topped the shard-1 wall-clock and removed a
meaningful rebalance lever — with the heaviest file locked to shard-1, the
top-N rule had no way to push any weight back the other direction as shard-2
grew. Splitting it into three cohesive files per tempo-architect's plan at
`C:/tmp/233-split-plan.md` restores that lever and slots cleanly into a new
N=4 shard-1 list.

Split (by cohesion, not by line range — titles preserved verbatim so
`--grep` patterns and CI summaries don't shift):

  test/session-phase-claim.test.ts       — 9 its (~22 s CI estimate)
      Claim/lease mechanics: booting→attached via claimAttachment, claim
      idempotency + AttachmentConflict + leaseMs validator, detached→attached
      re-attach, and the two heartbeat-as-renewal invariants (wrong-id silent
      drop + #119a leaseMs honored over the prior hardcoded default).

  test/session-phase-processing.test.ts  — 7 its (~18 s CI estimate)
      Processing/awaiting transitions from #117: attached→processing→awaiting
      via processingStart/End, supersession guard on processingStart, and the
      four awaiting-exit paths (→ processing, → draining, → detached via
      forceDetach, → gone via destroy).

  test/session-phase-detach.test.ts      — 9 its (~48 s CI estimate)
      requestDetach/drain/destroy paths: adapterExited graceful path, both
      #159 Gap 1 `skipTime` tests (drainingDeadline timeout + default 5s
      window — co-located so any future migration to `createTimeSkipping()`
      is one-file scope), forceDetach idempotency, destroy→gone terminal +
      WorkflowGone effects, setPreferredHost + orphanSummary round-trip,
      and expired-pre-seeded-attachment reap on boot.

Shared fixture: the 3-line `startFreshSession` helper is duplicated into
each new file per tempo-architect's recommendation — cheaper than a new
`test/phase-machine-fixtures.ts` surface, and avoids the "don't edit
helpers.ts during a rebalance" rule from test/README.md.

Shard rebalance (test/shard-config.json):

  Old N=3: session-phase-machine (88.7 s) + outbox (41.6 s) + scheduler
           (40.8 s) = 171.1 s / ~52% → drift 1.08× today, shrinking toward
           the 1.2× bound as more its accrue in the monolith.

  New N=4: session-phase-detach (~48 s) + outbox (41.6 s) + scheduler
           (40.8 s) + session-phase-claim (~22 s) = 152.4 s / ~46% →
           drift 1.17× against ~177 s shard-2. session-phase-processing
           (~18 s) slips to shard-2 as the lightest of the three.

Cross-references updated: `test/workflows/README.md` (bullet list + §"Adding
a new workflow-invariant test"), `test/conformance/conformance.test.ts`
(header docstring + two inline refs; line-number reference at :65 DROPPED
per the plan — line numbers shouldn't survive a split, the test title is
the stable handle), `test/session-lease.test.ts` (header docstring).
`test/README.md` §"2-way Mocha shard" updated to N=4 with the new file list.

Local wall-clocks (on this Windows host — absolute numbers slower than CI
but drift ratio is comparable):
  Shard-1: 215 s (52 tests passing)
  Shard-2: 224 s (794 passing / 22 pending)
  Drift:   1.04× — well inside the 1.2× rebalance bound.

All 25 tests from the monolith now live across the 3 new files and pass
individually; zero test-logic changes, pure file reorganization.

Verification per playbook: no `src/` touched, `npm run build` clean,
`scripts/check-test-ensemble-literals.sh` green, `git status` shows exactly
the expected 9-file diff (3 new + 1 delete + 2 cross-ref updates in code +
2 README updates + shard-config.json). Full `npm test` skipped — CI's
6-job matrix is authoritative.

Minor plan footnote: tempo-architect's plan header says "26 flat `it`
blocks" / detach cluster "10 tests"; actual count in the file is 25 its
with 9 in detach (one arithmetic slip in the plan — clustering itself is
unambiguous, so no deviation on split logic). Total 9+7+9=25 across the
new files.

Closes #233.
